Remove mirachSession and baseMirachSession (see below)
This may be the wrong move, but I felt that the refactor was being made more
complex by the multiple layers of oop on oop. I couldn't think of any outside
callers that would ever need to implement these structs / interfaces. In fact,
it seems like nothing other than mirach would ever implement that interface. I
considered its impact on privacy, but removing these does not seem to affect it.

It does significantly simplify the package. I believe it is easier to follow the
code path with fewer matryoshka dolls nested. It may make the cli and separating
into separate packages simpler. At least, that is the hope.